# rest-spec/typescript-client-generated/model/errorResponse.ts
/**
 * eDelphi REST API
 * REST API for eDelphi
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ErrorResponse { 
    message?: string;
}
export interface ErrorResponseOpt { 
    message?: string;
}